[PROMPT]: Add prompt versioning

+++ apps/prompt/backend/src/prompt/schemas.py
@@ -1,15 +1,25 @@
 """Data models."""
 
 # Standard Library
+import datetime
 import json
 from string import Formatter
-from typing import List, Optional, Union
+from typing import List, Optional
 
 # 3rd party libraries
 from pydantic import BaseModel
 
 # Source
+from src.prompt.exceptions import (
+    DeletionProtectedPrompt,
+    PromptNotFound,
+    PromptVersionNotFound,
+)
 from src.prompt.tables import PromptTemplateTable
+from src.settings import get_settings
+
+
+settings = get_settings()
 
 
 class PromptTemplateSchema(BaseModel):
@@ -22,7 +32,8 @@ class PromptTemplateSchema(BaseModel):
     id: Optional[str] = None
     template: str
     alias: str
-    created_at: Optional[str] = None
+    version: int = 0
+    created_at: Optional[datetime.datetime] = None
 
     @property
     def variables(self) -> List[str]:
@@ -40,6 +51,7 @@ class PromptTemplateSchema(BaseModel):
         prompt = PromptTemplateTable(
             template=self.template,
             alias=self.alias,
+            version=self.version,
         )
         prompt.save()
         prompt_dict = json.loads(prompt.to_json())
@@ -47,18 +59,39 @@ class PromptTemplateSchema(BaseModel):
             id=prompt_dict["id"],
             template=prompt_dict["template"],
             alias=prompt_dict["alias"],
+            version=prompt_dict["version"],
             created_at=prompt_dict["created_at"],
         )
 
     def delete(self) -> None:
-        prompt = PromptTemplateTable.get(self.alias)
-        prompt.delete()
+        """Deletes prompt from table."""
+        prompts = PromptTemplateTable.query(self.alias, scan_index_forward=False)
+
+        prompts = list(prompts)
+
+        if not prompts:
+            raise PromptNotFound(alias=self.alias)
+
+        for prompt in prompts:
+            for _, x in settings.LIST_OF_PROMPTS.items():
+                if prompt.alias in x[1]:
+                    raise DeletionProtectedPrompt(alias=self.alias)
+            prompt.delete()
 
     @classmethod
     def get(
-        cls, alias: Optional[str] = None
-    ) -> Union["PromptTemplateSchema", List["PromptTemplateSchema"]]:
-        """Returns row of the table."""
+        cls,
+        alias: Optional[str] = None,
+        version: Optional[int] = None,
+        raises_if_not_found: bool = False,
+    ) -> Optional[List["PromptTemplateSchema"]]:
+        """Returns row of the table.
+
+        Note: returns a list even when the
+
+        Raises:
+            PromptNotFound: of
+        """
         if alias is None:
             return list(
                 map(
@@ -66,16 +99,45 @@ class PromptTemplateSchema(BaseModel):
                     list(PromptTemplateTable.scan()),
                 )
             )
-        return PromptTemplateSchema(
-            **json.loads(PromptTemplateTable.get(alias).to_json())
+        if version is None:
+            # if no version specified get the latest.
+            query = PromptTemplateTable.query(alias, scan_index_forward=False)
+        else:
+            query = PromptTemplateTable.query(
+                alias, PromptTemplateTable.version == version
+            )
+
+        query = list(query)
+
+        if not query:
+            if raises_if_not_found:
+                if version is None:
+                    raise PromptNotFound(alias=alias)
+                else:
+                    raise PromptVersionNotFound(alias=alias, version=version)
+            return None
+
+        return list(
+            map(
+                lambda x: cls(
+                    id=x.id,
+                    template=x.template,
+                    created_at=x.created_at,
+                    version=x.version,
+                    alias=x.alias,
+                ),
+                query,
+            )
         )
 
-    def update(self, **kwargs) -> None:
-        """Updates table record."""
-        prompt = PromptTemplateTable.get(self.alias)
-        prompt.update(
-            actions=[PromptTemplateTable.template.set(kwargs.get("template"))]
-        )
+    def update(self, **kwargs) -> "PromptTemplateSchema":
+        """Updates table record from latest version."""
+        query = list(PromptTemplateTable.query(self.alias, scan_index_forward=False))
+        return PromptTemplateSchema(
+            template=kwargs.get("template"),
+            alias=self.alias,
+            version=int(query[0].version) + 1,
+        ).save()
 
 
 class PromptTemplateOutputSchema(BaseModel):
@@ -83,34 +145,28 @@ class PromptTemplateOutputSchema(BaseModel):
 
     id: Optional[str] = None
     template: str
-    created_at: Optional[str] = None
+    created_at: Optional[datetime.datetime] = None
     variables: List[str] = []
+    version: int
     alias: str
 
     @classmethod
     def from_template_schema(
-        cls, input: Union[PromptTemplateSchema, List[PromptTemplateSchema]]
-    ) -> "PromptTemplateOutputSchema":
+        cls, input: List[PromptTemplateSchema]
+    ) -> List["PromptTemplateOutputSchema"]:
         """Converts internal schema to output schema."""
-        if isinstance(input, list):
-            return list(
-                map(
-                    lambda x: cls(
-                        id=x.id,
-                        template=x.template,
-                        created_at=x.created_at,
-                        variables=x.variables,
-                        alias=x.alias,
-                    ),
-                    input,
-                )
+        return list(
+            map(
+                lambda x: cls(
+                    id=x.id,
+                    template=x.template,
+                    created_at=x.created_at,
+                    variables=x.variables,
+                    version=x.version,
+                    alias=x.alias,
+                ),
+                input,
             )
-        return cls(
-            id=input.id,
-            template=input.template,
-            created_at=input.created_at,
-            variables=input.variables,
-            alias=input.alias,
         )
